[POS-22] Add identity testing for serialization
This commit adds identity testing for Binary, MessagePack and SafeCopy
classes for most types in Pos.Types.Types.

It also bumps universum so that string conversion no longer uses
'toS' or the 'StringConv' class, relying instead on 'ConvertUtf8'
and more. It also removes usages of 'toS' and fixes errors resultant
from this change.