KademliaDHT: do not add closer on sendToNode
The closers are cleared only when the DHT is stopped. So remembering the
closer for each 'sendToNode' is a memory leak. It seems these closers
aren't necessary in practice, for when the DHT is stopped, the main
thread comes to an end and all threads will die anyway. If a set of
threads must be killed at a certain point, I recommend using a safer
abstraction like async.