[CSL-238] Refactor executables' CLI options
This commit refactors the CLI options in many of the project's
executables by moving them to Pos.CLI and grouping them with
general functions where possible.

The most numerous set of common arguments between all executables
was found, and moved to avoid redundancy.

Beware: since they are now exported as a single datatype,
`CommonArgs`, the order of the CLI otions has now changed in the
 `node`, `wallet` and  `smart-generator` executables.

Furthermore, when new CLI options are added that have:
* a `long` field
* a `metavar` field
* a `help` field
consider using `Pos.CLI.templateParser`.