[CSL-238] Fix error leftover from merge
This commit fixes an error caused by an unremoved merge marker,
and some sections of code added by it that were no longer in
the `master` branch.

Some warnings have also been removed, although most were left
as they were because they appear to be part of work in progress.

A force-push could have been done, but since this is the master
branch, the risk was not worth it. The commit history will have
this extraneous commit, but it's better to avoid messing up
our main development branch.