[CSL-421] Add serialization tests to block types
This commit adds serialization identity testing to Block
related types, as well as the necessary Arbitrary instances.

SafeCopy identity testing, namely, the Arbitrary instance for
Ssc ssc => Body (MainBlockchain ssc) has been fixed to take into
account the following invariant: in a block's body in the main
blockchain, the number of unique transactions in the merkle
tree and the number of transaction witness has to be the same.

Lastly, several warnings for redundant imports have been fixed.