[CSL-421] Refactor identity testing
This commit refactors identity testing so that the amount of
boilerplate is reduced - at least one line is still needed for
identity testing of a class like `Binary` or SafeCopy`, but in
nearly every case, it's shorter now.

Lastly, `Arbitrary` instances for `Block/BlockHeader` have
been removed because these types are only an `Either`
application on two types related to Genesis/Main blockchains,
and
`instance (Arbitrary a, Arbitrary b) => Arbitrary (Either a b)`
has already been defined in Test.QuickCheck.