No auto threading; remove time-warp dependencies
This is supposed to be a rework of time-warp, so it should not depend
upon time-warp.

Removed the 'RepeatForever' and 'SleepForever' notions because I'd like
to discuss them. I suspect they're just not necessary, because they can
be implemented in terms of other primitives.

The use of 'Node' has been changed so that it doesn't automatically
thread a list of workers. Instead, you get the 'Node m' and 'SendActions m'
and are free to do whatever 'm' actions you want. Once your action is
done, the 'EndPoint' created by the low-level node is closed and
everything shuts down.