Wait for handlers on close, improve error handling
It's possible that a SYN is received after an end point has closed. This
was observed in the test cases, in which 0 parcels were to be sent: the
sender would establish the connection but the end point would close
before the SYN was received. When it's received, the dispatcher tries to
ACK but it can't make a connection. Now the SYN will be ignored whenever
the end point has already closed.

Even with this fix, we would still observe an EventConnectionLost,
causing the dispatcher to blow up. But no more. Now an
EventConnectionLost is handled gracefully: all connections to that end
point have their ChannelIn plugged with a Nothing and no error is
thrown.