[CSL-421] Fix Relay types' serialization tests
This commit fixes binary serialization tests for the type
`DataMsg SlotHolderId GtMsgContents`.

It also updates the `Arbitrary` instance for `SlotId` because of
its new `Bi` instance, and removes a no longer necessary `NFData`
instance for `GoodTx` - it was used in a benchmark of transaction
verification which was later on removed.