Node tracks statistics about running handlers
The idea is that the amount of load induced by a peer is not necessarily
proportional to the rate of bytes coming in, so bounding the ingress buffers
isn't always enough. Another measure of load is the number of running handers
spawned as a result of connections from that peer. Perhaps there's a listener
for a very small message, but which is not cheap to run. The dispatcher must
be sensitive to this and gradually slow down service to a peer which hits this
listener very often. If some peer does so, then we should see the number of
remotely-initiated running handlers balloon far past the average over all
peers, and then impose a smaller buffer limit on input from that peer
(perhaps choosing 0 and only increasing it once the number of running
handlers has dwindled, although there's a risk of deadlock here in case
one of those handlers needs more input).

A node now tracks the following statistics, for remotely- and also
locally-initiated handlers which are currently running.

  - how many of each for a given peer
  - how many of each in total
  - the average of each over all peers
  - the average of the (square of each) over all peers, from which the
    standard deviation can be derived