ConversationActions doesn't give peerData
It was a mistake to have it in here, because it gives a duplicate in
case of a listener: a conversation listener gets the peer data as well
as conversation actions, which also supply the peer data.

The peer data was included in the conversation actions for the benefit
of 'withConnectionTo', so that an initiating peer can get the receiving
peer's data. This is still possible: the peer data is now supplied to
the continuation of 'withConnectionTo'.