Use EKG to monitor, custom nt, nt-tcp, tw-nt
The custom network-transport, network-transport-tcp, and time-warp-nt
branches chosen in stack.yaml fix some potentially very bad problems
with bidirectional connections: a handler can hang indefinitely in case
the connection is lost to the peer before the ACK is received. But
that's fixed now. The nt-tcp branch is also patched to deliver
ConnectionClosed events always, even if the connection is lost of the
endpoint is brought down. This makes it easier to clean up input
channels. The dispatcher remains defensive, though, and will plug input
channels on end point closed and connection lost events, even though it
shouldn't ever have to.