Really fix race in peer data sending
The previous commit fixed one side, but another remained.
The issue lies in either

  - Connecting when all established connections are in the process of
    closing.
  - Closing the last established connection while there are other
    connections in the process of being brought up.

If all established connections are in the process of closing, we must
wait for them all to close before attempting to establish a new one,
else we can't know whether to send the peer data.

If we try to close the last established connection, and there are other
being brought up, we must wait until one of them is established, or all
of them fail, before closing the last one, else we won't know whether to
send the peer data.

It's all about avoiding the "unstable" state, in which all connections
to that peer are either being brought up or being brought down.