Fill peer data exclusives for outbound handlers
This is to ensure that, if somehow the node goes down while the handlers
are running, or a connection is lost before peer data arrives, they
won't block when asking for peer data.

Also included better logging for threads that the dispatcher waits for.
And by using ThreadId and an assumption about its relation to the
promise (see notes inline) we have eliminated one extra thread creation
in spawnHandler.