Fix rare race in bidirectional connections
Fixed a race in establishing a bidirectional connection, which was
so unlikely to be observed in real world cases, but did come up when
using self-connections: it was possible for the ACK to be received
before the handler/nonce was registered in the node state. Now that's
solved: don't send the SYN until the nonce is registered.

Also improved testing.

Temporary: log 50 nt events on anomaly, timeouts

One of the CIs keeps showing anomalies (test hangs, or unknown
connections reported) but I can't reproduce these locally. This commit
should give more information in the CI log.