Fair QDisc implementations and simulation
Run the simulation like so:

  stack exec qdisc-simulation -- <qdisc> <duration>

where qdisc is one of
  "fair", "fair_stm", "one_place", "unbounded"
and duration parses to a whole number, determining for how many seconds
to run the simulation.
Yes there are two implementations of the fair QDisc. Seems they have
very close performance characteristics. Maybe we should favour the STM
variant because it's simpler.
You can change the distributions for reader and writers by modifying
Simulation.hs. What's in there now shows a case in which the fair QDisc
really is fairer than the one-place.