[CSL-728] Add tests to 'computeSharesDistr'
This commit covers the 'computeSharesDistr' function with tests,
making some necessary changes to the codebase.

The contents of the
'Toss' submodule in 'Pos/Ssc/GodTossing' were not being exported,
so it was added to the re-export list in 'Pos/Ssc/GodTossing.hs'.

However, this caused conflicts with the functions 'hasOpening',
'hasCommitment' and 'hasShares' from 'Pos.Ssc.GodTossing.Functions'.

To solve this, their counterparts from 'Pos.Ssc.GodTossing.Toss.Base'
were renamed.