Use finer-grained MVars for PeerStatistics.
Previously, we had a map holding the PeerStatistics for all peers,
requiring us to hold the MVar of the Statistics for the node whenever we
wanted to update statistics for a peer.  Since we're tracking in-flight
bytes, this could lead to contention of that MVar.

Now, each PeerStatistics is held in a dedicated MVar, removing this
possibility of contention.