Add ReadSharedAtomic constructor
It's possible that reading a shared atomic can be implemented in a way
that is somewhat faster / has less contention with concurrent uses. Now
we allow for it. The Production interpretation is also updated to use
`readMVar`, which as of base 4.9.1.0 is a GHC primitive, not implemented
in terms of `takeMVar`/`putMVar`.