[CSL-854] Improve code afret PR review
The 'BadOpening' type was extraneous as an opening is never invalid
by itself, only when associated with a mismatching commitment.
As such, it has been removed.

Moreover, testing revealed 'verifyCommitment' would successfully
verify an invalid 'Commitment' which had an empty 'commShares' field,
so this check has been added to that verifying function.