[CSL-921] Keep nodes from entering recovery mode (#138)
* Additional debug messages for block relay

* More debug messages.

* Do not enter recovery when receiving new block.

Previously, addToBlockRequestQueue would set ncRecoveryHeader if it was
not previously set, causing the node to enter recovery mode.  In
particular, this would happen each time that handleUnsolicitedHeader got
a header that continued the current chain.

This commit fixes that behaviour.

* Remove excessive debug messages

While tracking down why nodes kept ending up in recovery mode, I added
some debug messages, but they will probably not be useful in the future
and only add to the noise, so I'm removing them again.

Also, removed a comment.