[TW-134] 0.0.0.0 is neither default not valid host
A node must not advertise its external address as 0.0.0.0.
First, because the host at which it is reached is assumed by time-warp
to match the host at which it claims to be reachable (the one in the
EndPointAddress).
Second, because with an upcoming patch to network-transport-tcp, any
EndPoint's reported host must match the host which its peer determines
from the TCP connection. This is a security measure, to prevent an easy
DOS attack where any node can deny service to another by claiming the
other's host.