[CSL-865] Cosmetics
This commit changes the names of the 'SharedSecrets' record fields
to be in line with the style guide.

It also removes redundant parens, and uses the 'RecordWildCards'
extension, which is activated by default in the entire project,
instead of the 'NamedFieldPuns' which needlessly increases code
 verbosity.

Lastly, instances of 'join (f <$> m)' are replaced by the equivalent
'm >>= f'.