[CSL-870] Support unaddressable nodes
Wallet and smart-generator give Nothing for bpIpPort. This means their
transports will be unaddressable rather than claim 127.0.0.1 as their
host.

Unfortunately, Kademlia will still run, so its peers will see a bogus
node in the network, and we can expect full nodes to regularly time-out
while trying to connect to these. That fix is for another day.

The peer-host check is also enabled on the TCP transport. This means
that running the wallet or smart-generator as addressable while behind
NAT will cause these programs to fail to connect, which is good because
accepting their claimed 127.0.0.1 host will only cause problems.