[CSL-620] Add tests to 'modifyPollModifier'
Some properties for 'Pos.Update.Poll.Modifier.modifyPollModifier'
have been added.

In the course of writing these properties, a bug with the 'Monoid
 PollModifier' instance was found, as this type with the previous
function does not form a monoid, but a semigroup. There is no
identity because the 'pmEpochProposers' field always defaults to
the second argument's, meaning it is not conserved when applied
with an old modifier to a new, empty poll modifier.

Moreover, there already was a property to check that a type with
a certain operation was a group, so some refactoring was done
to avoid duplication, both now and in the future.

In 'Test.Pos.Util' there are now properties for up to a commutative
monoid, with more complex algebraic requirements being added in the
test module itself if necessary.