[CSL-993] Sort out work with balances (aka stakes)
Now there are two versions of functions which get balances from DB with
different names: with `effective` word in name and with `real` word in name.
Real corresponds to stake which is modified by transactions and is stored
in DB.
Effective is same is real if we are not in bootstrap era, otherwise it's taken
from predefined constants.
This commit also fixes a bug where effective stake was used is 'MonadBalances'
implementation.