[TW-139] Correct dispatcher exception throwing / handling (#39)
* Re-throw dispatcher exceptions

A fairly egregious oversight by me! The application needs to know if the
dispatcher stops doing its thing.

* Throw exception if EndPoint closed prematurely

* Adjust exception throwing in dispatcher

DO throw exceptions on inconsistent internal state.

DO NOT throw an exception in case the connection is lost to a peer which
is not feeding an application handler. This was a mistake, I can't
imagine why I thought it was the right thing to do...