Emergency NAT patch
Bumps nt-tcp and time-warp revisions so that peers which claim 0.0.0.0
or 127.0.0.1 as their hosts will be treated as unaddressable.

This mitigates the self-inflicted denial-of-service observed when many
preconfigured wallet programs try to talk to the network, each claiming
the very same address. nt-tcp is designed to admit at most one TCP
socket per pair of EndPoints, which are identified by their addresses.

This is a temporary patch, which will make existing client executables
play well with the network. It should be obsoleted soon by a major
update. The revision of nt-tcp which is chosen will certainly not be
admitted upstream.