Fix AppVeyor artifact path (#259)
* Fix AppVeyor artifact path

Since build 0.4.4258 the artifact zip hasn't included all the
assets. When paths were restructured to account for Windows' path
length limit, the artifact target path was missed.

* AppVeyor: copy artifact assets back to repo dir

According to the AppVeyor docs, "the artifact path must be relative to
the root of repository." Using the absolute path to the custom
%WORK_DIR% won't work. The strategy here is to copy assets from the
work dir back to where they would have been if we built in the default
project dir (we don't due to Windows' path length restriction).