[CSL-980] Add tests to 'PurePoll' behavior
This commit adds two tests, one trivial and the other not, to the
behavior of the 'PurePoll' monad that was designed to simulate the
update system's functionality.

In the first test, invariance of the order in which poll modifiers
are applied to each other and then to a 'PollState', or each separately
to the 'PollState', is verified.

In the second, a list of random 'MonadPoll' actions is created, which
is then applied to a 'PurePoll' in sequence. Finally, the state is
extracted from the monad and compared to the state that would be obtained
by applying to it each modifier to which each 'MonadPoll' action
is equivalent to. They must be equal.