[CSL-980] Fix 'PollModifier' test
Previously, a CI failure led to the discovery of a problem with a
test in 'PollSpec', and after debugging, the cause was uncovered.

Essentially, in the 'modifyPollState' function, the
'_psActivePropsIdx' field of the 'PollState' wasn't being correctly
modified.

Even after fixing it to remove the proposals marked for deletion in
'pmDelActivePropsIdx', there was still an issue as the proposals
marked for addition in 'pmActiveProps' were also being neglected.

After adding that logic to the function, the failing test was still
not successful.

After some effort to rule out the test being incorrect by temporarily
disabling the 'InsertActiveProposal' field of the 'PollAction' type,
which caused the test to pass, the problem was narrowed down to the
'insertActiveProposal' method in the 'MonadPoll PurePoll' instance.
It was adding the proposal to 'psActiveProposals', but it wasn't
adding it, along with its corresponding application name as its key,
to 'psActivePropsIdx'. After fixing the instance, the test passed.
To fix this logical error in the 'deactivateProposal' method, the
proposal being deleted is also removed from 'psActivePropsIdx'.

On the technical side, the 'Data.HashMap.Strict.differenceWith'
function has been added to subtract hashmaps with regards to their
values in the 'modifyPollState' function, as for every application
name key in 'psActivePropsIdx', it subtracts from its 'HashSet UpId'
the 'HashSet' of update proposal ids that 'pmDelActivePropsIdx'
is to remove from the state. This function is only available in the
'unordered-containers'' API starting with version 0.2.8.0. Due to
some issues with dependency solving if this version is requested of
all instances of this package in 'cardano-sl.cabal', it has
temporarily been added to 'stack.yaml' as an 'extra-dep'.