Avoid excessive retries in modifyTxpLocalData
Neil observed that modifying the TxpLocalData is performed in one atomic
STM transaction, that can potentially take a non-significant amount of
time.  Concurrent attempts can thus trigger endless retries and block
the system from making progress.

This could explain an observation in the benchmarks, where a single node
would not record any transactions.

This commit tests the hypothesis, by putting a simple semaphor in front
of the STM action in modifyTxpLocalData.  It's not meant to be
production code.