[CSL-1004] Upgrade to latest `binary`.
This fixes unnecessary ByteString copies with `BS.append`
(see https://github.com/kolmodin/binary/pull/115)
which causes allocations to be reduced by ~30% in cardano-sl.

Note since `binary` is a package that ghc depends on,
cardano-sl will not be able to use the `ghc` library
until the next GHC release (8.2), but it currently doesn't
use the `ghc` library so that's fine.