[CSL-1004] Enable TCP_NODELAY, enforce TCP_QUICKACK on Linux.
This removes delays in TCP networking and thus makes node catchup
slightly faster.

Note that TCP_QUICKACK only works on Linux, and only on
Linux >= 2.4.4.
On other operating systems other ways to disable delayed acknowledgements
will be needed (for example, there's a registry entry on Windows).

For background on the two network options, see:

* https://eklitzke.org/the-caveats-of-tcp-nodelay
* http://jvns.ca/blog/2015/11/21/why-you-should-understand-a-little-about-tcp/

For background on why they should be used with network-transport-tcp, see:

* https://github.com/haskell-distributed/network-transport-tcp/issues/66

For the upstream PR that adds TCP_QUICKACK, see:

* https://github.com/haskell-distributed/network-transport-tcp/pull/67

Note that while at the time of writing the network-transport-tcp docs say
that TCP_NODELAY is on by default, this is not true:

* https://github.com/haskell-distributed/network-transport-tcp/issues/62