[CSM-242] Fix hanging tests
This commit restores the test suite to its original speed, which
will prevent CI from failing over hanging tests.

The issue was introduced in commit '999b0fb0' on April 26th, 2017.
The reasons are still unknown and require further investigation,
but changing the length of the empty value of the 'PassPhrase'
type from the 'Pos.Crypto.SafeSigning' module from '0' (which is
the length of 'mempty :: ScrubbedBytes') to
'32 = length $ replicate 32 0' caused a tenfold slowdown in all
cryptographic primitive tests, indicating less efficiency in
types from 'Pos.Crypto' involving the 'PassPhrase' type, which in
turn caused the whole test suite to break the CI's build limit.

In addition to this, and unrelated to CSM-242, some linting was
performed to allow CI to pass.