MemPoolModifyReason, includes allocations
The MemPoolModifyReason type is used in place of string. The FromJSON is
not completely backwards compatible: a text reason will be ignored and
defaulted to Unknown. I suspect this is OK since we have little or no
data with text reasons, but I could fix it if necessary.

Bytes allocated is included in each mempool event. These are the
allocations during the STM transaction.