[TW-138] provide peer data on all interactions (#54)
* [TW-138] peer data available on send

This seems to be the simplest, most most noninvasive solution to the
problem of having peer data avilable at all times when interacting with
a peer. Previously, it wasn't available for single-message interactions,
so single-message interactions/listeners have been removed. Such a thing
can be expressed as a conversation where one of the send/recv types is
Void. Everything just works out for free: since the conversation must be
ACKd by the peer, and since the peer data is always delivered by the
first connection from a peer, it will always be available when the
conversation initiator or listener runs.

The benchmarks have probably been messed up. They compile, but it's not
clear whether they still do what they need to do. Better benchmarking is
required anyway.

Although this is simple and noninvasive, I do not believe it's the right
way forward. The problem is that the programmer must commit to doing a
single conversation based on the peer data, whereas the value of the
peer data might reasonably suggest a multiple-conversation interaction.
So that's it: the current design doesn't support this notion of a
session, in which a peer can be contacted and its peer data held across
0 or more message-based conversations.

* [TW-138] listeners depend on peer data

Now you can choose an entirely different set of listeners depending upon
the peer data, rather than having the set of listeners fixed and
modulating each one based on the peer data.

* [TW-138] peer data available on send

This seems to be the simplest, most most noninvasive solution to the
problem of having peer data avilable at all times when interacting with
a peer. Previously, it wasn't available for single-message interactions,
so single-message interactions/listeners have been removed. Such a thing
can be expressed as a conversation where one of the send/recv types is
Void. Everything just works out for free: since the conversation must be
ACKd by the peer, and since the peer data is always delivered by the
first connection from a peer, it will always be available when the
conversation initiator or listener runs.

The benchmarks have probably been messed up. They compile, but it's not
clear whether they still do what they need to do. Better benchmarking is
required anyway.

Although this is simple and noninvasive, I do not believe it's the right
way forward. The problem is that the programmer must commit to doing a
single conversation based on the peer data, whereas the value of the
peer data might reasonably suggest a multiple-conversation interaction.
So that's it: the current design doesn't support this notion of a
session, in which a peer can be contacted and its peer data held across
0 or more message-based conversations.

* [TW-138] listeners depend on peer data

Now you can choose an entirely different set of listeners depending upon
the peer data, rather than having the set of listeners fixed and
modulating each one based on the peer data.

* Fix bench warnings and errors