[TW-149] Redesign of message serialization (#64)
* Redesign of message serialization

- Outgoing messages are split into chunks of at most some number of
  bytes. This value must be properly configured when running over
  network-transport-tcp, to be smaller than the maximum receive size.
- Node.Message has been split up and the Packable/Unpackable classes
  rolled into one class Serializable, which demands an incremental
  decoder which mimics the on from Data.Binary.Get
- In a conversation, 'recv' takes a limit on the number of bytes which
  will be read.
- Can use store instead of binary for deserialization, as shown in the
  implementation of Node.Message.Store

* Correct send framing

* Naming and comments, no program changes

* Update comment on 'recv'

* Fix test builds

* Fix benchmarks builds

* MTU is configurable in NodeEnvironment

* Send peer data respecting mtu, correct writeMany

* Immediately use trailing bytes from peer data

If the peer data and a complete handshake are sent all at once, the
dispatacher wouldn't digest and respond to the handshake until the
sender sends more data, possibly leading to deadlock.
This has been fixed.

* Fix tests for the benefit of CI

Using TCP_NODELAY makes things much faster. The MTU chosen in the tests
is not ridiculously small, as that would also makes things slower. Tests
are run to only 10 successes. The CI machines should now be able to do
them before timing out.