[TW-148] import relay stuff from CSL (#62)
* Provide a configurable delay on incomming connections

We have an existing configurable delay on all incomming
network-transport level messages. This adds a configurable delay on just
incomming lightweight connections. So if this delay is used and the
general one is set to no delay, then we would only get delays on new
lightweight connections. This corresponds to high level / app level
conversations, which in the current app are short-lived. So this delay
more closely corresponds to high level messages.

* [TW-148] import relay stuff from CSL

* Update the examples for the node init API change

* [TW-148] Fix tests after merge of master

* [TW-148] fix bench after merge of master

* [TW-148] fix after rebase on master TW-149