[CSM-307] Fix `importWallet`
This makes two things:

1. `changeWalletPass` endpoint does nothing if pass doesn't change.
It's important, since it creates accounts which correspond to new pass,
then sends money to them and removes new accounts. In case when new pass is same
as old pass, new accounts are also the same and everything goes bad.

2. Now, on pass change, old accounts are not totally removed from db,
but just marked as _removed_ - that's needed for correct transaction history
fetching.