[TW-148] Use time-warp relay queueing
There's a --node-type command-line argument (relay, core, or edge) which
defaults to NodeCore. Core and relay peers can also be specified on the
command line to give the classification. Peers which don't appear in
this classification are assumed to be edge nodes. It's cumbersome to
specify the configuration like this so soon we'll parse it from a yaml
file.

Major change: 'SendActions' is not longer defined in time-warp, nor is
the abstract 'OutboundQueue' interface.

This removes converseToNeighbors