[CSL-1290] Add BlockVersionModifier
Previously 'BlockVersionData' was misused. It represented two different
things:
• description of the currently used block version;
• modification of the currently used block version proposed in update.

Eveyrthing was fine as long as we wanted to represent modification as the
whole 'BlockVersionData'. But recently we added tx fee policy which can be
potentially big and decided to put it into update proposal only if it differs
from the currently adopted one. It lead to necessity to separate
'BlockVersionData' and 'BlockVersionModifier' which was done in this PR.

Also now we can easily make other proposed fields optional.