[CSL-1161] Refactor 'Arbitrary' instances
This commit refactors 'Arbitrary' instances from the 'src/' folder.
That means the instances in 'Pos.Block.Arbitrary',
'Pos.Delegation.Arbitrary' and 'Pos.Txp.Network.Arbitrary' were
changed.

Furthermore, an artifact from a previous merge (most likely #981)
was removed. An 'Arbitrary' module that is no longer used was kept
in the project, so to prevent conflicts and the compilation of an
unnecessary module it was deleted.