[CSL-1161] Refactor 'Arbitrary' instances in 'update/'
This commit's message will serve as an explanation for all the
previous ones concerning CSL-1161.

This issue is about writing better 'Arbitrary' instances by using
generics to remove the applicative pattern
'Constructor <$> arbitrary <*> arbitrary <*> ...' used often when
writing these class instances.

Moreover, the 'shrink' method of this class was seldom written
(that's mostly my fault), which leads to less useful information
both in quantity and quality when tests fail, as QuickCheck is
not able to reduce the failing value to any special or meaningful
edge case (the default implementation is 'shrink = []').

As such, and again relying on generics when possible to prevent
redundancy, this method was added to most 'Arbitrary' instances
around the project.