[CSL-1162] Add missing serialization tests to 'src/' types
This commit message shall also serve as explanation for the
previous 4 commits.

Over the course of development, types are added, refactored or
deleted and have their 'Bi'nary serialization instances
written/changed/removed accordingly.

Because ensuring these instances are well tested is critical for
proper functioning of the code, an effort has to be made from
time to time to update these tests to reflect the codebase's
recent changes. This is what the issue CSL-1162 is about.

This commit in particular updates binary testing for types in the
'src/' folder.