[CSL-1161] Fix custom integral types' 'Arbitrary' instances
This commit changes the 'Arbitrary' instances for 'UnsignedVarInt',
'SIgnedVarInt' and 'FixedSizeInt' to use the strategy given by the
'GeneralizedNewTypeDeriving' extension. Using 'shrinkIntegral' here
is not ideal, because the underlying type may use a more complicated
shrinking method and this should be preserved for the varint types.

This commit message rewords a comment by Ivan Gromakovskii (@gromak)
made on the PR.