[CSL-1181] Refactor heavyweight delegation to use StakeholderIds mostly
It's needed to implement CSL-1181 which is about removing PSKs from the nodes
which stopped being ricmhen. The reason it's needed is that we store only
'StakeholderId's of richmen, so we can't delete data associated with
'PublicKey's.

Another advantage is that 'StakeholderId' is small than 'PublicKey', so
disk space consumption becomes smaller.