Add initial instances for Pos.Core.Fee
This commit adds the initial `Bi` instances for the data types in
Pos.Core.Fee. On top of that, it extends the `Pos.Binary.Cbor.Test`
module with some basic roundtrip property for `Coeff`, as interim
for more properties for the more interesting data types in the module.

Last but not least, it removes the `autoexporter` pragma from the
unmbrella-module `Pos.Binary.Cbor` as it was importing `Test` by
default thus leading to an import cycle.

Signed-off-by: Alfredo Di Napoli <alfredo.dinapoli@iohk.io>