Use mapConcurrently for launching plugins
The Mockable.Concurrent doesn't provide mapConcurrently_ but that would
be a little nicer.

The message has been changed to announcing that we're launching n
plugins. If any of the plugins throw an exception we'll find out.

The runWallet function waits for any of the plugins to fail. If
runWallet is killed, then those child threads will be killed
automatically too.