Fix a subtle bug in the `Ser` macro.
This commit fixes a subtle bug in the `Pos.Crypto.AsBinary.Ser` macro definition
Specifically, the auto-generated instance was not idempotent, as it was
calling `encode/serialize'` in the `fromBinary` implementation, thus
adding more headers (in CBOR-jargon) to the resulting object.

Instead, we use `getAsBinary` to get the BS blob and proceed to the
reification into a proper `a` object.