Add ability to enqueue to a specific set of nodes (#80)
* Add ability to enqueue to a specific set of nodes

This set of nodes is still subject to the enqueueing policy. This PR
also adds a new kind of message type called `MsgRequestData`. This new
message type, along with the `enqueueTo` set of API calls, can be used
when requesting data (such as blocks) from peers that are known to have
them. The advantage is that the queue can then make an educated guess of
_which_ of the peers-known-to-have-the-data we should request the data
from; for example, we prefer to ask data from relay nodes over asking
data from core nodes so that we don't put more load on the system when
there are nodes in recovery mode.

* Introduce MsgRequestBlockHeaders message

also rename `MsgBlock` to `MsgAnnounceBlockHeader`, and rename
`MsgRequestData` to `MsgRequestBlock`.

* Add enqueueAsync and enqueueAsyncTo

* [TW-148] async enqueue is prime

Others are defined in terms of it, and importantly, the OutboundQueue
uses it.

* [TW-148] fix mistake in demo from last commit