[TW-144] ignore incoming data if handler finished (#55)
This prevents a simple attack (described in the ticket) in which anybody
who knows what peer data you expect can fill up your heap with bytes.

Note that there's another side to the story in case this is backed by
network-transport-tcp: the transport's QDisc. If the simple unbounded
QDisc is used, the attack will probably still work, because the
unbounded Chan will fill up with data faster than the dispatcher can
remove it. Using a one-place or fair QDisc mitigates this.