Add `--init-delay` command line arg
This can be used to delay initialization of the node until a certain
time (but after the node has initialized its network interface, so that
other nodes can connect to it).

This is useful to avoid the network layer deciding that it cannot reach
other nodes (because they have not yet finished booting up) and putting
them on in failure mode, and subsequently refusing to send messages to
them for a while.