[CSL-1483] Change 'Address' type according to the new structure
This commit changes 'Address' type according to the description from
CSL-1483. For more details see CSLREQ-201.
It mostly consists of necessary refactoring to make code compile.
However, there are some other changes, specifically:
1. 'Address' serialization changed (obviously).
2. 'AddressIgnoringAttributes' newtype was removed.
3. I also deleted logic implicitly assigning all stake to destination
stakeholder if explicit distribution is empty and address is PubKey,
because it's impossible to extract 'StakeholderId' from an 'Address'.