[CSL-1490] eliminate get block deadlock
The maximum in-flight conversations, as given by the default dequeue
policy, for a relay to a core, is 1. When in recovery mode, a relay
requests headers and then, within that same conversation, attempts to
pull in the blocks by enqueueing a MsgGetBlocks. That's all fine except
that it waits, within that conversation, on the results of the
MsgGetBlocks conversation(s). That's the deadlock: those MsgGetBlocks
conversation(s) push the in flight over 1 and so are never dequeued and
the requestHeaders conversation never finishes.

This patch has us enqueue the MsgGetBlocks and not wait on the results.
I'm not sure if this breaks any block logic assumptions, but at the very
least it solves the deadlock.