[CSL-1483] Include explicit dsitribution in txs
The logic implicitly assigning all stake to one stakeholder if
distribution is empty has been removed, so now it's necessary to always include
explicit distribution.

I did it for SendToAll command in lwallet, because it seems to be the only
case where we don't have bootstrap era. I also added some comments.
After [CSL-1489] it will be unncessary, because distribution from addrses
will be used instead.
I also did it for block payload generator, because it's easy and will be
needed later.

I also had to get rid of Text.Parsec.String in favor of Text.Parsec.Text
for consistency. Text version was chosen because we prefer 'Text' wherever
possible.