[CSL-1438] Fix UtxoSpec tests
These tests got broken because we were parsing error messages which is
bad in general and fragile.
We decided to continue effort of changing 'verifyTxUtxo' to use
'MonadError ToilVerFailure'. I partially did it, but then gave up and decided
to leave it to @neongreen who told he could do it tomorrow.
Currently we just check that type of error is expected, but don't check any
details.